third_party/WebKit/LayoutTests/fast/css/font-face-style-attribute.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<style>
@font-face {
  font-family: Font1;
  src: url(font.ttf);
  font-weight: bold;
}
</style>

<script>

var fontStyle = document.styleSheets[0].cssRules[0].style;
var originalFontCSSText = fontStyle.cssText;

test(() => {
  assert_equals(fontStyle.cssText,
      "font-family: Font1; src: url(\"font.ttf\"); font-weight: bold;");
}, "FontFace style serializes correctly");

test(() => {
  var cssText = "font-family: Font2; src: url(\"font2.ttf\");";
  fontStyle.cssText = cssText;
  assert_equals(fontStyle.cssText, cssText);

  fontStyle.cssText = originalFontCSSText;
}, "FontFace.style's cssText can be set");

test(() => {
  assert_equals(fontStyle.length, 3);
}, "FontFace.style length is retrieved");

test(() => {
  assert_equals(fontStyle[0], "font-family");
  assert_equals(fontStyle[1], "src");
  assert_equals(fontStyle[2], "font-weight");
}, "FontFace.style attributes are in order");

test(() => {
  assert_equals(fontStyle.getPropertyValue("font-family"), "Font1");
  assert_equals(fontStyle.getPropertyValue("src"), "url(\"font.ttf\")");
  assert_equals(fontStyle.getPropertyValue("font-weight"), "bold");
}, "GetPropertyValue returns correct values");

test(() => {
  assert_equals(fontStyle.getPropertyPriority("font-family"), "");
  assert_equals(fontStyle.getPropertyPriority("src"), "");
  assert_equals(fontStyle.getPropertyPriority("font-weight"), "");
}, "GetPropertyPriority returns empty string for everything");

test(() => {
  // Property already on fontStyle
  fontStyle.setProperty("font-weight", "100");
  // Property not on fontStyle
  fontStyle.setProperty("font-style", "italic");
  // Attribute style setter
  fontStyle.fontStretch = "condensed";

  var cssText = "font-family: Font1; src: url(\"font.ttf\"); " +
    "font-stretch: condensed; font-weight: 100; font-style: italic;";
  assert_equals(fontStyle.cssText, cssText);

  // Reset.
  fontStyle.cssText = originalFontCSSText;
}, "SetProperty sets the property");

test(() => {
  fontStyle.setProperty("--my-property", "100");

  assert_equals(fontStyle.cssText, originalFontCSSText);
}, "SetProperty has no effect for custom properties");

test(() => {
  fontStyle.removeProperty("font-weight");
  assert_equals(fontStyle.cssText,
      "font-family: Font1; src: url(\"font.ttf\");");

  // Reset.
  fontStyle.cssText = originalFontCSSText;
}, "RemoveProperty removes the property");

test(() => {
  assert_equals(fontStyle.cssFloat, "");

  fontStyle.cssFloat = 5;
  assert_equals(fontStyle.cssFloat, "");
}, "Accessing the .cssFloat attribute has no effect and does not crash");

</script>
